refactor(settings): import COOKIE_NAME through the settings barrel
